+++ cli/pom.xml
@@ -25,7 +25,7 @@
         </dependency>
         <dependency>
             <groupId>io.github.belgif.rest.guide.validator</groupId>
-            <artifactId>belgif-rest-guide-validator-rules</artifactId>
+            <artifactId>belgif-rest-guide-validator-runner</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
