--- core/src/main/java/io/github/belgif/rest/guide/validator/output/OutputType.java
@@ -1,8 +1,0 @@
-package io.github.belgif.rest.guide.validator.output;
-
-/**
- * Output type for OpenApiValidator.
- */
-public enum OutputType {
-    CONSOLE, JUNIT, LOG4J, JSON, NONE
-}
