--- core/src/main/java/io/github/belgif/rest/guide/validator/output/model/ViolationEntry.java
@@ -1,11 +1,0 @@
-package io.github.belgif.rest.guide.validator.output.model;
-
-public record ViolationEntry(
-        String ruleId,
-        String description,
-        String message,
-        String level,
-        String fileName,
-        int lineNumber,
-        String pointer) {
-}
