--- core/src/main/java/io/github/belgif/rest/guide/validator/output/model/ViolationGroup.java
@@ -1,8 +1,0 @@
-package io.github.belgif.rest.guide.validator.output.model;
-
-import java.util.List;
-
-public record ViolationGroup(
-        int total,
-        List<ViolationEntry> occurrences) {
-}
