+++ runner/src/main/java/io/github/belgif/rest/guide/validator/runner/output/OutputType.java
@@ -1,0 +1,8 @@
+package io.github.belgif.rest.guide.validator.runner.output;
+
+/**
+ * Output type for OpenApiValidator.
+ */
+public enum OutputType {
+    CONSOLE, JUNIT, LOG4J, JSON, NONE
+}
