+++ runner/src/main/java/io/github/belgif/rest/guide/validator/runner/output/model/ViolationGroup.java
@@ -1,0 +1,8 @@
+package io.github.belgif.rest.guide.validator.runner.output.model;
+
+import java.util.List;
+
+public record ViolationGroup(
+        int total,
+        List<ViolationEntry> occurrences) {
+}
